add styling to new task form
the new task form is now within a horizontally centered container that inhabits 2/3 of the horizontal space in the viewport.

form labels are aligned on the left side of this container, and form fields are aligned on the right side of it.

there is some padding to vertically separate each of the form label+field rows.

the submit button is horizontally centered.